log: don't use snprintf when no formatting is required.
This saves an expensive call to snprintf and also allows larger (static)
string sizes to be logged.

+++ application.c
@@ -1380,11 +1380,11 @@ static void wlan_scan_done_callback(void *arg, STATUS status)
 		"UNKNOWN",
 	};
 
-	log("wlan scan result: %s\n", status <= CANCEL ? status_msg[status] : "<invalid>");
-	log("> %-16s %-4s %-4s %-18s %-13s %-13s %-6s %s\n", "SSID", "CHAN", "RSSI", "AUTH", "PAIR CIPHER", "GROUP_CIPHER", "OFFSET", "BSSID");
+	logf("wlan scan result: %s\n", status <= CANCEL ? status_msg[status] : "<invalid>");
+	logf("> %-16s %-4s %-4s %-18s %-13s %-13s %-6s %s\n", "SSID", "CHAN", "RSSI", "AUTH", "PAIR CIPHER", "GROUP_CIPHER", "OFFSET", "BSSID");
 
 	for(bss = arg; bss; bss = bss->next.stqe_next)
-		log("> %-16s %4u %4d %-18s %-13s %-13s %6d %02x:%02x:%02x:%02x:%02x:%02x\n",
+		logf("> %-16s %4u %4d %-18s %-13s %-13s %6d %02x:%02x:%02x:%02x:%02x:%02x\n",
 				bss->ssid,
 				bss->channel,
 				bss->rssi,
