+++ display_ssd1306.c
@@ -587,7 +587,7 @@ bool display_ssd1306_picture_load(unsigned int picture_load_index)
 
 	if(string_size(&flash_sector_buffer) < SPI_FLASH_SEC_SIZE)
 	{
-		log("display ssd1306: load picture: sector buffer too small: %u\n", flash_sector_buffer_use);
+		logf("display ssd1306: load picture: sector buffer too small: %u\n", flash_sector_buffer_use);
 		return(false);
 	}
 
@@ -611,7 +611,7 @@ bool display_ssd1306_layer_select(unsigned int layer)
 
 	if((flash_sector_buffer_use != fsb_free) && (flash_sector_buffer_use != fsb_config_cache))
 	{
-		log("display ssd1306: load picture: flash buffer not free, used by: %u\n", flash_sector_buffer_use);
+		logf("display ssd1306: load picture: flash buffer not free, used by: %u\n", flash_sector_buffer_use);
 		return(false);
 	}
 
@@ -619,7 +619,7 @@ bool display_ssd1306_layer_select(unsigned int layer)
 
 	if(spi_flash_read(display_picture_load_flash_sector * SPI_FLASH_SEC_SIZE, string_buffer_nonconst(&flash_sector_buffer), SPI_FLASH_SEC_SIZE) != SPI_FLASH_RESULT_OK)
 	{
-		log("display ssd1306: load picture: failed to read sector: 0x%x\n", display_picture_load_flash_sector);
+		logf("display ssd1306: load picture: failed to read sector: 0x%x\n", display_picture_load_flash_sector);
 		goto error;
 	}
 
@@ -627,7 +627,7 @@ bool display_ssd1306_layer_select(unsigned int layer)
 
 	if(!string_match_cstr(&flash_sector_buffer, pbm_header))
 	{
-		log("display ssd1306: show picture: invalid image header: %s\n", string_to_cstr(&flash_sector_buffer));
+		logf("display ssd1306: show picture: invalid image header: %s\n", string_to_cstr(&flash_sector_buffer));
 		goto error;
 	}
